+++ edx_proctoring/tests/test_handlers.py
@@ -4,7 +4,6 @@ Tests for handlers.py
 from unittest.mock import patch
 
 import ddt
-from httmock import HTTMock
 
 from django.db.models.signals import pre_delete, pre_save
 
@@ -38,14 +37,6 @@ class SignalTests(ProctoredExamTestCase):
         super().tearDown()
         pre_delete.disconnect()
         pre_save.disconnect()
-
-    @patch('logging.Logger.error')
-    def test_backend_fails_to_delete_attempt(self, logger_mock):
-        # If there is no response from the backend, assert that it is logged correctly
-        with HTTMock(None):
-            self.attempt.delete_exam_attempt()
-            log_format_string = 'Failed to remove attempt_id=%s from backend=%s'
-            logger_mock.assert_any_call(log_format_string, 1, self.backend_name)
 
     @ddt.data(None, MockInstructorService())
     @patch('edx_proctoring.handlers.get_runtime_service')